Redux Actions.
We made an action object in the ActionCreator.
comments.js is a reducer function that is connected to the Configure Store.
The configure store get stimulated via the dispatch function. Which is implemented in the MainComponent as mapDispatchToProps.

So an action is created via addComment. This calls the store.dispatch(). Which calls all the reducer functions (Since we make use of CombineReducers all the reducer fns are called when a state is changed)
This is why we implement a switch statement to switch in the action.type (in comments.js; todo dishes.js -and the other 2.)

This reducer functions take an Action and the previous state as inputs. And returns a mutable object.

IMPORTANT: Reducers: i.e: comment.js
Do not mutate the state, meaning: We cannot modify the state we are sent in as a paramter. We can add to it if we want and return the modified version as a new object.
Look at the concat example